requirements.txt
numpy==1.19.4
opencv-python-headless==4.5.1.48
pandas==1.3.1
Pillow==9.3.0
streamlit==1.14.0
sklearn==0.0
